Fix clippy lints

+++ src/ec.rs
@@ -467,7 +467,7 @@ impl WriterBase<WriterEncoder> {
     let mut c = 0;
     let mut offs = self.s.precarry.len();
     // dynamic allocation: grows during encode
-    let mut out = vec![0 as u8; offs];
+    let mut out = vec![0_u8; offs];
     while offs > 0 {
       offs -= 1;
       c += self.s.precarry[offs];
